+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - (Base) New `DeckPane` component with swipe and slide transition support.
 - (Base) New `MaskTextField` (and `MaskTextFormatter`) component to support masked text input.
+- (Base) New `PasswordTextField` component to simplify `PasswordTextFormatter` usage.
 - (CSS) 🚀 New MacOS-like Cupertino theme in light and dark variants.
 - (CSS) 🚀 New [Dracula](https://ui.draculatheme.com/) theme.
 - (CSS) New `TabPane` style. There are three styles supported: default, floating and classic (new one).